review: final code quality pass for SmolLM2 (#25)
* review: code quality + correctness test fixes

- Remove GPU softmax import from _sample() (numpy softmax, no GPU roundtrip)
- Remove stale double-normalization in softmax
- Add decode mode guard: reject seq>1 when cache is non-empty
- Fix download script docstring
- Fix correctness test: properly track decode token agreement

* fix: softmax float64 + decode_pass from Gemini review

---------

+++ smollm2_triton/model.py
@@ -450,7 +450,7 @@ class SmolLM2ForCausalLM:
             scaled = logits / temperature
             # NumPy softmax
             logits_stable = scaled - np.max(scaled)
-            probs = np.exp(logits_stable)
+            probs = np.exp(logits_stable).astype(np.float64)
             probs = probs / np.sum(probs)
             if top_k > 0:
                 top_k = min(top_k, len(probs))
